formatter for any source
